[feature/#928] activity -> screen
